+++ tiny_cheetah/tui/chat_menu.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 import os
+import threading
 import time
 from pathlib import Path
 from typing import Dict, List, Optional
@@ -10,21 +11,12 @@ import tinygrad as tg
 from transformers import AutoTokenizer
 from rich.markup import escape
 
-# Some macOS Python builds hit `bad value(s) in fds_to_keep` when hf_transfer is enabled.
-# Disable it here so the downloader falls back to the pure-python implementation.
-os.environ.setdefault("HF_HUB_ENABLE_HF_TRANSFER", "0")
-
 from tiny_cheetah.models.llm.helpers import generate, load_safetensors
 from tiny_cheetah.models.llm.model import Model
 from tiny_cheetah.models.llm.model_config import ModelConfig
 from tiny_cheetah.models.llm.shard import Shard
-
-try:  # pragma: no cover - optional dependency
-    import huggingface_hub
-    from tiny_cheetah.repos import RepoCustom
-except Exception:  # pragma: no cover - missing huggingface_hub
-    huggingface_hub = None  # type: ignore[assignment]
-    RepoCustom = None  # type: ignore[assignment]
+from tiny_cheetah.tui.widget.model_picker_screen import ModelPickerScreen
+from tiny_cheetah.repos import RepoCustom
 
 from textual.app import ComposeResult
 from textual.binding import Binding
@@ -37,8 +29,6 @@ from textual.widgets import Button, Footer, Header, Input, Label, RichLog, Stati
 
 
 class ChatModelSelected(Message):
-    """Message emitted when the user selects a new model."""
-
     def __init__(self, sender: MessagePump, model_id: str) -> None:
         super().__init__(sender)
         self.model_id = model_id
@@ -69,6 +59,7 @@ class ChatScreen(Screen[None]):
         self._model_cache_path: Optional[Path] = None
         self._history: List[Dict[str, str]] = []
         self._generation_thread_in_progress = False
+        self.out_tokens: List[int] = []
 
     def compose(self) -> ComposeResult:
         yield Header(show_clock=True)
@@ -119,10 +110,11 @@ class ChatScreen(Screen[None]):
             if self._model is not None and self._tokenizer is not None:
                 self._append_system("Model already loaded.")
                 return
-            self._append_system(f"Loading model '{self._model_id}'...")
-            time.sleep(1)  # allow UI to update
+            
             try:
-                self._load_model_blocking()
+                load_model_t = threading.Thread(target=self._load_model_thread, args=(self,))
+                load_model_t.start()
+                load_model_t.join()
             except Exception as exc:  # pragma: no cover
                 self._append_system(f"Failed to load model: {exc}")
         elif event.button.id == "chat-back":
@@ -164,7 +156,9 @@ class ChatScreen(Screen[None]):
         if self._model is None or self._tokenizer is None:
             self._append_system("Loading model now...")
             try:
-                self._load_model_blocking()
+                load_model_t = threading.Thread(target=self._load_model_thread, args=(self,))
+                load_model_t.start()
+                load_model_t.join()
             except Exception as exc:  # pragma: no cover
                 self._append_system(f"Failed to load model: {exc}")
                 return
@@ -201,7 +195,19 @@ class ChatScreen(Screen[None]):
                 f"({datetime.now().strftime('%Y-%m-%d %H:%M:%S')}) [bold][#ff0000]{label}[/][/bold]: {text}\n"
             )
 
-    def _load_model_blocking(self) -> None:
+    @staticmethod
+    def _config_get(config: object, key: str, default=None):
+        if config is None:
+            return default
+        if isinstance(config, dict):
+            return config.get(key, default)
+        if hasattr(config, "__getitem__"):
+            value = config[key]  # type: ignore[index]
+            return default if value is None else value
+        return default
+
+    def _load_model_thread(self) -> None:
+        self._append_system(f"Loading model '{self._model_id}'...")
         if not self._model_id:
             raise RuntimeError("No model selected")
         start = time.time()
@@ -215,17 +221,6 @@ class ChatScreen(Screen[None]):
         self._model_cache_path = model_path
         self._append_system(f"Model ready in {elapsed:.1f}s.")
 
-    @staticmethod
-    def _config_get(config: object, key: str, default=None):
-        if config is None:
-            return default
-        if isinstance(config, dict):
-            return config.get(key, default)
-        if hasattr(config, "__getitem__"):
-            value = config[key]  # type: ignore[index]
-            return default if value is None else value
-        return default
-
     def _load_local_config(self, directory: Path) -> tuple[Path, ModelConfig]:
         config = ModelConfig()
         config_file = directory / "config.json"
@@ -236,6 +231,44 @@ class ChatScreen(Screen[None]):
         if gen_config.exists():
             config.load_generation_config(gen_config)
         return directory, config
+
+    def _load_model_worker(self):
+        log_messages: List[str] = []
+        candidate_path = Path(self._model_id).expanduser()
+        if candidate_path.exists():
+            log_messages.append(f"Loading local model from {candidate_path}")
+            model_path, model_config = self._load_local_config(candidate_path)
+            tokenizer_source = str(model_path)
+            tokenizer_local = True
+        else:
+            log_messages.append(f"Downloading model from Hugging Face: {self._model_id}")
+            if RepoCustom is None:
+                raise RuntimeError("huggingface_hub is required to download models.")
+            repo = RepoCustom(self._model_id)
+            model_path, model_config, repo_messages = repo.download()
+            log_messages.extend(f"[download] {msg}" for msg in repo_messages)
+            log_messages.append(f"Model cached at {model_path}")
+            tokenizer_source = str(model_path)
+            tokenizer_local = True
+
+        num_layers = int(self._config_get(model_config, "num_layers", 0))
+        shard = Shard(
+            self._model_id,
+            start_layer=0,
+            end_layer=num_layers,
+            total_layers=num_layers + 1,
+        )
+        log_messages.append(f"Instantiating shard {shard}")
+        model = Model(model_config, shard)
+        load_safetensors(
+            model,
+            model_path,
+            model_config,
+            weight_device=os.getenv("TC_DEVICE", "CPU"),
+            use_tied=bool(self._config_get(model_config, "tie_word_embeddings", False)),
+        )
+        tokenizer = AutoTokenizer.from_pretrained(tokenizer_source, local_files_only=tokenizer_local)
+        return model, model_config, tokenizer, model_path, log_messages
 
     def _generate_response(self, max_new_tokens: int = 128) -> None:
         if self._model is None or self._tokenizer is None:
@@ -310,43 +343,7 @@ class ChatScreen(Screen[None]):
         self._append_system("Model cleared. Select and load a model to continue.")
 
 
-    def _load_model_worker(self):
-        log_messages: List[str] = []
-        candidate_path = Path(self._model_id).expanduser()
-        if candidate_path.exists():
-            log_messages.append(f"Loading local model from {candidate_path}")
-            model_path, model_config = self._load_local_config(candidate_path)
-            tokenizer_source = str(model_path)
-            tokenizer_local = True
-        else:
-            log_messages.append(f"Downloading model from Hugging Face: {self._model_id}")
-            if RepoCustom is None:
-                raise RuntimeError("huggingface_hub is required to download models.")
-            repo = RepoCustom(self._model_id)
-            model_path, model_config, repo_messages = repo.download()
-            log_messages.extend(f"[download] {msg}" for msg in repo_messages)
-            log_messages.append(f"Model cached at {model_path}")
-            tokenizer_source = str(model_path)
-            tokenizer_local = True
-
-        num_layers = int(self._config_get(model_config, "num_layers", 0))
-        shard = Shard(
-            self._model_id,
-            start_layer=0,
-            end_layer=num_layers,
-            total_layers=num_layers + 1,
-        )
-        log_messages.append(f"Instantiating shard {shard}")
-        model = Model(model_config, shard)
-        load_safetensors(
-            model,
-            model_path,
-            model_config,
-            weight_device=os.getenv("TC_DEVICE", "CPU"),
-            use_tied=bool(self._config_get(model_config, "tie_word_embeddings", False)),
-        )
-        tokenizer = AutoTokenizer.from_pretrained(tokenizer_source, local_files_only=tokenizer_local)
-        return model, model_config, tokenizer, model_path, log_messages
+   
 
     def _generate_tokens_worker(
         self,
@@ -374,60 +371,3 @@ class ChatScreen(Screen[None]):
         )
         elapsed = time.time() - start
         return out_tokens, elapsed
-
-
-class ModelPickerScreen(ModalScreen[str | None]):
-    """Modal popup allowing the user to enter a model identifier or path."""
-
-    CSS_PATH = Path(__file__).with_name("chat_menu.tcss")
-
-    BINDINGS = [
-        Binding("escape", "dismiss(None)", "Cancel"),
-    ]
-
-    def __init__(self, initial_value: str) -> None:
-        super().__init__(id="model-picker")
-        self._initial = initial_value
-        self._input: Optional[Input] = None
-        self._cache_options = self._discover_cached_models()
-
-    def compose(self) -> ComposeResult:
-        with Container(id="model-picker-container"):
-            yield Static("Select Model", id="model-picker-title")
-            yield Label("Enter a Hugging Face repo or choose a cached model:", id="model-picker-help")
-            if self._cache_options:
-                with Container(id="model-picker-cache"):
-                    yield Label("Cached Models", id="model-picker-cache-title")
-                    for option in self._cache_options:
-                        yield Button(option, classes="model-cache-option")
-            input_field = Input(id="model-picker-input")
-            self._input = input_field
-            yield input_field
-            with Container(id="model-picker-buttons"):
-                yield Button("Cancel", id="model-picker-cancel")
-                yield Button("Apply", id="model-picker-apply", variant="primary")
-
-    def on_mount(self, _: Mount) -> None:
-        if self._input is not None:
-            self._input.value = self._initial
-            self.call_after_refresh(self._input.focus)
-
-    def on_button_pressed(self, event: Button.Pressed) -> None:
-        if event.button.id == "model-picker-cancel":
-            self.dismiss(None)
-        elif event.button.id == "model-picker-apply":
-            self.dismiss(self._input.value.strip() if self._input else None)
-        elif "model-cache-option" in event.button.classes:
-            self.dismiss(event.button.label.plain)
-
-    def _discover_cached_models(self) -> list[str]:
-        if huggingface_hub is None:
-            return []
-        cache_dir = Path(huggingface_hub.constants.HF_HUB_CACHE).expanduser()
-        if not cache_dir.exists():
-            return []
-        options: set[str] = set()
-        for repo_dir in cache_dir.glob("models--*"):
-            resolved = repo_dir.name.replace("models--", "").replace("--", "/")
-            options.add(resolved)
-        return sorted(options)
