+++ tiny_cheetah/tui/chat_menu.tcss
@@ -21,6 +21,8 @@ Screen {
     background: #050505;
     color: $text;
     padding: 1;
+    text-wrap: wrap;
+    overflow-y: auto;
 }
 
 .rich-log-line.system {
@@ -77,68 +79,3 @@ Footer {
     dock: bottom;
 }
 
-/* Modal styling */
-ModalScreen#model-picker {
-    align: center middle;
-}
-
-#model-picker-container {
-    width: 60%;
-    max-width: 80%;
-    border: round $accent;
-    background: $surface;
-    padding: 2;
-    layout: vertical;
-}
-
-#model-picker-title {
-    text-style: bold;
-    text-align: center;
-    margin-bottom: 1;
-}
-
-#model-picker-help {
-    color: $text-muted;
-    margin-bottom: 1;
-}
-
-#model-picker-cache {
-    border: round $accent;
-    padding: 1;
-    margin-bottom: 1;
-    layout: vertical;
-    height: 10;
-    overflow-y: auto;
-}
-
-#model-picker-cache-title {
-    text-style: bold;
-    margin-bottom: 1;
-}
-
-.model-cache-option {
-    margin-bottom: 1;
-    text-align: left;
-}
-
-.model-cache-option:last-child {
-    margin-bottom: 0;
-}
-
-#model-picker-buttons {
-    layout: horizontal;
-    align: right middle;
-    margin-top: 1;
-}
-
-#model-picker-buttons Button#model-picker-apply {
-    margin-left: 1;
-}
-
-#actions-panel Button {
-    margin-top: 1;
-}
-
-#actions-panel Button:first-child {
-    margin-top: 0;
-}
